Remove redundant extension
