+++ package.json
@@ -34,7 +34,6 @@
     "@stackla/lambda-api-bootstrap": "^1.0.12",
     "@types/cookie-parser": "^1.4.7",
     "@types/glidejs__glide": "^3.6.5",
-    "@types/isotope-layout": "^3.0.14",
     "@types/jest": "^29.5.13",
     "@typescript-eslint/eslint-plugin": "^7.13.1",
     "chokidar": "^4.0.1",
@@ -72,7 +71,6 @@
     "cookie-parser": "^1.4.6",
     "happy-dom": "^14.12.3",
     "hbs": "^4.2.0",
-    "isotope-layout": "^3.0.6",
     "serverless-hooks-plugin": "^1.1.0",
     "serverless-http": "^3.2.0",
     "swiper": "^11.1.14"
