+++ widgets/grid/layout.hbs
@@ -1,9 +1,9 @@
 
-<masonry-layout class="ugc-tiles" cols="auto" gap="10">
+<div class="ugc-tiles grid">
     {{#tiles}}
         {{>tpl-tile}}
     {{/tiles}}
-</masonry-layout>
+</div>
 <div id="buttons">
     <a id="load-more">Load More</a>
 </div>
