+++ widgets/grid/tile.hbs
@@ -1,7 +1,7 @@
-{{#tile data-hash=@index}}
-<div class="tile">
+{{#tile data-hash=@index class="grid-item"}}
+<div class="tile" data-background-image="{{image_thumbnail_url}}">
   <div class="tile-loading"></div>
-  <div class="tile-content hidden">
+  <div class="tile-content">
     {{#each attrs}}
       {{#ifequals this "instagram.reel"}}
         <div class="content-icon icon-reel"></div>
@@ -21,10 +21,6 @@
     {{/ifequals}}
     <span class="tile-timephrase">{{time_ago}}</span>
     <shopspot-icon tile-id={{id}} />
-  </div>
-  <div class="tile-image">
-    {{#lazy image_thumbnail_url "120"}}{{/lazy}}
-    <div class="swiper-lazy-preloader"></div>
   </div>
 </div>
 {{/tile}}
