+++ widgets/grid/widget.settings.ts
@@ -1,6 +1,5 @@
 import { Sdk } from "@stackla/ugc-widgets"
 import getCSSVariables from "@widgets/libs/css-variables"
-import { MasonryLayout } from "@appnest/masonry-layout"
 import {
   loadTitle,
   addAutoAddTileFeature,
@@ -24,31 +23,4 @@ export function loadWidgetSettings() {
   loadExpandedTileFeature(widgetSettings, onTileExpand, onTileClosed, onTileRendered)
   addTilesPerPageFeature(widgetSettings)
   addLoadMoreButtonFeature(widgetSettings)
-
-  const tileSizes: { [key: string]: string } = {
-    small: "127.8px",
-    medium: "210.4px",
-    large: "265.5px"
-  }
-
-  const refreshMasonryLayout = () => {
-    const masonryLayout = sdk.querySelector<MasonryLayout>("masonry-layout")
-    if (masonryLayout instanceof MasonryLayout) {
-      masonryLayout.layout()
-      if (widgetSettings.margin) {
-        masonryLayout.setAttribute("gap", `${widgetSettings.margin}`)
-      }
-      if (widgetSettings.tile_size) {
-        masonryLayout.setAttribute("maxColWidth", tileSizes[widgetSettings.tile_size])
-      }
-    }
-  }
-
-  sdk.addEventListener("moreLoad", () => {
-    refreshMasonryLayout()
-  })
-  window.addEventListener("scroll", () => {
-    refreshMasonryLayout()
-  })
-  sdk.addEventListener("tilesUpdated", refreshMasonryLayout)
 }
