+++ widgets/masonry/widget.settings.ts
@@ -10,8 +10,11 @@ import {
 import { addCSSVariablesToPlacement } from "@widgets/libs/widget.layout"
 import { getConfig } from "./widget.config"
 import { onTileExpand, onTileClosed, onTileRendered } from "@widgets/libs/extensions/swiper/swiper.expanded-tile"
-import { refreshMasonryLayout, reinitialiseMasonryLayout } from "@widgets/libs/extensions/masonry.extension"
-import { resizeAllUgcTiles } from "./masonry.lib"
+import {
+  refreshMasonryLayout,
+  reinitialiseMasonryLayout,
+  resizeAllUgcTiles
+} from "@widgets/libs/extensions/masonry.extension"
 
 declare const sdk: Sdk
